#341 DC0006 ProcedureRequiresDocumentation (Internal)
