+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalHasDiagnostic/CodeunitAccessInternal.al
@@ -1,0 +1,8 @@
+codeunit 50100 MyCodeunit
+{
+    Access = Internal;
+
+    procedure [|MyProcedure|]()
+    begin
+    end;
+}
