+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalHasDiagnostic/CodeunitAccessInternalProcedureWithAttribute.al
@@ -1,0 +1,9 @@
+codeunit 50100 MyCodeunit
+{
+    Access = Internal;
+
+    [NonDebuggable]
+    procedure [|MyProcedure|]()
+    begin
+    end;
+}
