+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalHasDiagnostic/CodeunitAccessInternalProcedureWithComment.al
@@ -1,0 +1,9 @@
+codeunit 50100 MyCodeunit
+{
+    Access = Internal;
+
+    // This is a comment
+    procedure [|MyProcedure|]()
+    begin
+    end;
+}
