+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalHasDiagnostic/InternalProcedure.al
@@ -1,0 +1,6 @@
+codeunit 50100 MyCodeunit
+{
+    internal procedure [|MyProcedure|]()
+    begin
+    end;
+}
