+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalHasDiagnostic/InternalProcedureWithAttribute.al
@@ -1,0 +1,7 @@
+codeunit 50100 MyCodeunit
+{
+    [NonDebuggable]
+    internal procedure [|MyProcedure|]()
+    begin
+    end;
+}
