+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalHasDiagnostic/InternalProcedureWithComment.al
@@ -1,0 +1,7 @@
+codeunit 50100 MyCodeunit
+{
+    // This is a comment
+    internal procedure [|MyProcedure|]()
+    begin
+    end;
+}
