+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalNoDiagnostic/CodeunitAccessInternalProcedureDocumentationComment.al
@@ -1,0 +1,12 @@
+codeunit 50100 MyCodeunit
+{
+    Access = Internal;
+
+    /// <summary>
+    /// This method...
+    /// </summary>
+    procedure [|MyProcedure|]()
+    begin
+
+    end;
+}
