+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalNoDiagnostic/CodeunitAccessInternalProcedureDocumentationCommentWithAttribute.al
@@ -1,0 +1,13 @@
+codeunit 50100 MyCodeunit
+{
+    Access = Internal;
+
+    /// <summary>
+    /// Sets the call body content from a stream.
+    /// </summary>
+    /// <param name="ContentStream">The instream containing the call body.</param>
+    [NonDebuggable]
+    procedure [|MyProcedure|](ContentStream: InStream)
+    begin
+    end;
+}
