+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalNoDiagnostic/CodeunitAccessInternalProcedureDocumentationCommentWithMultipleAttributes.al
@@ -1,0 +1,13 @@
+codeunit 50100 MyCodeunit
+{
+    Access = Internal;
+
+    /// <summary>
+    /// My procedure.
+    /// </summary>
+    [NonDebuggable]
+    [Scope('OnPrem')]
+    procedure [|MyProcedure|]()
+    begin
+    end;
+}
