+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalNoDiagnostic/Procedure.al
@@ -1,0 +1,6 @@
+codeunit 50100 MyCodeunit
+{
+    procedure [|MyProcedure|]()
+    begin
+    end;
+}
