+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalNoDiagnostic/ProcedureLocal.al
@@ -1,0 +1,6 @@
+codeunit 50100 MyCodeunit
+{
+    local procedure [|MyProcedure|]()
+    begin
+    end;
+}
