+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/InternalNoDiagnostic/TestCodeunit.al
@@ -1,0 +1,9 @@
+codeunit 50100 MyTestCodeunit
+{
+    Subtype = Test;
+
+    [Test]
+    procedure [|MyTestProcedure|]()
+    begin
+    end;
+}
