+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/PublicHasDiagnostic/Procedure.al
