+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/PublicHasDiagnostic/ProcedureWithAttribute.al
