+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/PublicHasDiagnostic/ProcedureWithComment.al
