+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/PublicNoDiagnostic/CodeunitAccessInternal.al
