+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/PublicNoDiagnostic/ProcedureDocumentationCommentWithAttribute.al
