+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/PublicNoDiagnostic/ProcedureInternal.al
