+++ src/ALCops.DocumentationCop.Test/Rules/PublicProcedureRequiresDocumentation/PublicNoDiagnostic/TestCodeunitHandlerMethod.al
